+++ stream-access-control/access-control-core/src/test/java/com/backbase/stream/service/AccessGroupServiceTest.java
@@ -60,6 +60,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import com.backbase.stream.product.task.BatchProductGroupTask;
+import com.backbase.stream.product.task.BatchProductIngestionMode;
 import com.backbase.stream.worker.exception.StreamTaskException;
 import com.backbase.stream.worker.model.StreamTask;
 import lombok.AllArgsConstructor;
@@ -341,7 +342,7 @@ class AccessGroupServiceTest {
         BatchProductGroupTask batchProductGroupTask = new BatchProductGroupTask().data(
             new BatchProductGroup().serviceAgreement(new ServiceAgreement().externalId("sa_benedict").internalId("sa-internal-id"))
         );
-        batchProductGroupTask.setIngestionMode(BatchProductGroupTask.IngestionMode.UPDATE);
+        batchProductGroupTask.setIngestionMode(BatchProductIngestionMode.upsert());
 
         Map<BusinessFunctionGroup, List<BaseProductGroup>> baseProductGroupMap = new HashMap<>();
         baseProductGroupMap.put(new BusinessFunctionGroup().id("business-function-group-id-1"), Collections.emptyList());
@@ -408,7 +409,7 @@ class AccessGroupServiceTest {
         BatchProductGroupTask batchProductGroupTask = new BatchProductGroupTask().data(
             new BatchProductGroup().serviceAgreement(new ServiceAgreement().externalId("sa_benedict").internalId("sa-internal-id"))
         );
-        batchProductGroupTask.setIngestionMode(BatchProductGroupTask.IngestionMode.UPDATE);
+        batchProductGroupTask.setIngestionMode(BatchProductIngestionMode.upsert());
 
         Map<BusinessFunctionGroup, List<BaseProductGroup>> baseProductGroupMap = new HashMap<>();
         baseProductGroupMap.put(
@@ -480,7 +481,7 @@ class AccessGroupServiceTest {
         BatchProductGroupTask batchProductGroupTask = new BatchProductGroupTask().data(
             new BatchProductGroup().serviceAgreement(new ServiceAgreement().externalId("sa_benedict").internalId("sa-internal-id"))
         );
-        batchProductGroupTask.setIngestionMode(BatchProductGroupTask.IngestionMode.REPLACE);
+        batchProductGroupTask.setIngestionMode(BatchProductIngestionMode.replace());
 
         Map<BusinessFunctionGroup, List<BaseProductGroup>> baseProductGroupMap = new HashMap<>();
         baseProductGroupMap.put(new BusinessFunctionGroup().id("business-function-group-id-1"), Collections.emptyList());
@@ -528,7 +529,7 @@ class AccessGroupServiceTest {
         BatchProductGroupTask batchProductGroupTask = new BatchProductGroupTask().data(
             new BatchProductGroup().serviceAgreement(new ServiceAgreement().externalId("sa_benedict").internalId("sa-internal-id"))
         );
-        batchProductGroupTask.setIngestionMode(BatchProductGroupTask.IngestionMode.UPDATE);
+        batchProductGroupTask.setIngestionMode(BatchProductIngestionMode.upsert());
 
         Map<BusinessFunctionGroup, List<BaseProductGroup>> baseProductGroupMap = new HashMap<>();
         baseProductGroupMap.put(new BusinessFunctionGroup().id("business-function-group-id-1"), Collections.emptyList());
@@ -585,7 +586,7 @@ class AccessGroupServiceTest {
         BatchProductGroupTask batchProductGroupTask = new BatchProductGroupTask().data(
             new BatchProductGroup().serviceAgreement(new ServiceAgreement().externalId("sa_benedict").internalId("sa-internal-id"))
         );
-        batchProductGroupTask.setIngestionMode(BatchProductGroupTask.IngestionMode.UPDATE);
+        batchProductGroupTask.setIngestionMode(BatchProductIngestionMode.upsert());
 
         Map<BusinessFunctionGroup, List<BaseProductGroup>> baseProductGroupMap = new HashMap<>();
         baseProductGroupMap.put(