+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.88.0](https://github.com/Backbase/stream-services/compare/2.87.0...2.88.0)
+### Changed
+- Replaced `BatchProductGroupTask.IngestionMode` with more flexible `BatchProductIngestionMode`. New class keeps ingestion modes separately for each main resource
+involved in BatchProductIngestionSaga processing: function groups, data groups and arrangements. Two preset modes have been created: `BatchProductIngestionMode.upsert` 
+and `BatchProductIngestionMode.replace` (equivalent of previous `UPDATE` and `REPLACE`, respectively), but new ones can be composed of any combination of "sub modes".
+
 ## [2.86.1](https://github.com/Backbase/stream-services/compare/2.86.0...2.86.1)
 ### Fixed
 - Added qualifier for WebClient in ContactsServiceConfiguration.